+++ sinp_nomenclatures/urls.py
@@ -1,37 +1,25 @@
 from django.urls import path
 
 from .views import (
-    ItemViewset,
+    NomenclatureViewset,
     SourceViewset,
     TypeViewset,
 )
-
+from django.urls import path, include
+from rest_framework.routers import DefaultRouter
 
 app_name = "sinp_nomenclatures"
 
+router = DefaultRouter()
+
+router.register(r'sources', SourceViewset,'sources')
+router.register(r'types', TypeViewset, 'types')
+router.register(r'nomenclatures', NomenclatureViewset, 'nomenclaturess')
+
 urlpatterns = [
+    path('', include(router.urls)),
     path(
-        "nomenclature/sources",
-        SourceViewset.as_view({"get": "list"}),
-        name="source_list",
-    ),
-    path(
-        "nomenclature/source/<int:pk>",
-        SourceViewset.as_view({"get": "retrieve"}),
-        name="source",
-    ),
-    path(
-        "nomenclature/types",
-        TypeViewset.as_view({"get": "list"}),
-        name="type_list",
-    ),
-    path(
-        "nomenclature/type/<int:pk>",
-        TypeViewset.as_view({"get": "retrieve"}),
-        name="type",
-    ),
-    path(
-        "nomenclature/type/<int:pk>/items",
+        "nomenclatures/type/<int:pk>/item",
         TypeViewset.as_view({"get": "list"}),
         name="item_list",
     ),