+++ docs/changelog.rst
@@ -2,6 +2,16 @@
  CHANGELOG
 ===========
 
+v0.1.1
+======
+
+Update poetry config
+
+v0.1.1
+======
+
+* Rename "Item" model to "Nomenclature"
+
 v0.1.0
 ======